Remove POSIX specific ssize_t usage
Related-issues: APPLINK-22713, APPLINK-22716

+++ src/components/utils/test/auto_trace_test.cc
@@ -107,7 +107,7 @@ typedef std::pair<std::string, std::string> LogWithLevel;
  * @return index of first element from search_for vector or -1 if all elements
  * exists in stream
  */
-ssize_t CheckIfIstreamContains(std::ifstream& stream,
+int32_t CheckIfIstreamContains(std::ifstream& stream,
                                const std::vector<LogWithLevel>& search_for) {
   std::vector<LogWithLevel>::const_iterator it = search_for.begin();
   std::string line;
@@ -139,7 +139,7 @@ TEST(AutoTraceTest, DISABLED_AutoTrace_WriteToFile_ReadCorrectString) {
   log_messages.push_back(LogWithLevel(trace_log_level, enter_message));
   log_messages.push_back(LogWithLevel(debug_log_level, debug_message));
   log_messages.push_back(LogWithLevel(trace_log_level, exit_message));
-  const ssize_t index_of_missed_element =
+  const int32_t index_of_missed_element =
       CheckIfIstreamContains(file_log, log_messages);
   if (index_of_missed_element != -1) {
     const std::string missed_log = log_messages[index_of_missed_element].first +
